pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "uFocus"
version = "2.3.2"
dynamic = ["readme"]
authors = [
    {name = "Dimitrios Papaioannou", email = "dimipapaioan@outlook.com"},
]
description = "uFocus: An autofocusing system for the nuclear microprobe at Uppsala University's Tandem Laboratory"
dependencies = [
    "pyside6-qtads == 4.3.1.4",
    "pyqtgraph",
    "opencv-python-headless",
    "pypylon",
    "pyserial",
    "numpy",
    "scipy",
]
requires-python = ">=3.10"
license = {text = "MIT License"}
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT License",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
]

[project.scripts]
ufocus = "ufocus.main:run_main"

[tool.setuptools.dynamic]
readme = {file = "README.md"}

[tool.ruff]
target-version = "py310"
src = ["ufocus"]

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I"]
